+++ CHANGELOG.md
@@ -3,13 +3,16 @@
 
 ## [ver. 3.57] - xx.09.2024
 ## UPDATE
-- UPDATE - [Portraits]: Background setting is now always available.
-- UPDATE - [Portraits]: Update Class Icons for the new Version of JiberishUI Icons.
 - UPDATE - [Important-Spells]: Default Texture function.
+- UPDATE - [Portraits]: Background setting is now always available.
+- UPDATE - [Portraits]: Optimized Style c and b.
+- UPDATE - [Portraits]: Update Class Icons for the new Version of JiberishUI Icons.
 ### FIX 
+- FIX - [Portraits]: Smal Bugs with ne Textures.
 - FIX - [Portraits]: Old DB check did not worked correctly.
 ## NEW
 - NEW - [Portraits]: Add Rare/ Elite/ Boss to custom Texture section.
+- NEW - [Portraits]: Add simple Dragon texture to Circle, Drop & Pure.
 - NEW - [Portraits]: New color function for Rare/ Elite/ Boss textures.
 
 ## [ver. 3.56] - 02.09.2024