+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - UPDATE - [Important-Spells]: Default Texture function.
 - UPDATE - [Portraits]: Background setting is now always available.
 - UPDATE - [Portraits]: Optimized Style c and b.
-- UPDATE - [Portraits]: Update Class Icons for the new Version of JiberishUI Icons.
+- UPDATE - [Portraits]: Update Class Icons for the new Version of JiberishUI Icons - Thx Repooc.
 ### FIX 
 - FIX - [Portraits]: Smal Bugs with ne Textures.
 - FIX - [Portraits]: Old DB check did not worked correctly.